stateless-comm-reporter/requirements.txt
fastapi==0.112.0
uvicorn==0.30.5
langfuse==2.39.2
python-dotenv==1.0.1
google-cloud-aiplatform==1.59.0
google-cloud-documentai==2.29.2
google-auth==2.32.0
google-api-python-client==2.129.0
firebase-admin==6.5.0
networkx==3.3

-e graph2nosql  
-e graphrag_lite


